Only use trylock not to lock ourselves out
With pthread_mutex_lock we would sometimes get into a deadlock by the
combination of lock for the vector (timed_waits_mutex) and the
passed_lock. By using a trylock we can try to notify all threads of
timeouts but postpone it to the next time update if we fail to do so.